requirements.txt
anyio==3.6.2
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
deta==1.1.0
fastapi==0.92.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.2
newspaper3k==0.2.8
nltk==3.8.1
Pillow==9.4.0
pycodestyle==2.10.0
pydantic==1.10.5
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.25.0
tinysegmenter==0.3
tldextract==3.4.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14